@@ -58,7 +58,7 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   /// The returned [Future] completes after listeners have been notified.
   Future<void> updateMapOptions(
     Map<String, dynamic> optionsUpdate, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('updateMapOptions() has not been implemented.');
   }
@@ -71,7 +71,7 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   /// The returned [Future] completes after listeners have been notified.
   Future<void> updateMarkers(
     MarkerUpdates markerUpdates, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('updateMarkers() has not been implemented.');
   }
@@ -84,7 +84,7 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   /// The returned [Future] completes after listeners have been notified.
   Future<void> updatePolygons(
     PolygonUpdates polygonUpdates, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('updatePolygons() has not been implemented.');
   }
@@ -97,7 +97,7 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   /// The returned [Future] completes after listeners have been notified.
   Future<void> updatePolylines(
     PolylineUpdates polylineUpdates, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('updatePolylines() has not been implemented.');
   }
@@ -110,7 +110,7 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   /// The returned [Future] completes after listeners have been notified.
   Future<void> updateCircles(
     CircleUpdates circleUpdates, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('updateCircles() has not been implemented.');
   }
@@ -122,8 +122,8 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   ///
   /// The returned [Future] completes after listeners have been notified.
   Future<void> updateTileOverlays({
-    Set<TileOverlay> newTileOverlays,
-    @required int mapId,
+    Set<TileOverlay>? newTileOverlays,
+    required int mapId,
   }) {
     throw UnimplementedError('updateTileOverlays() has not been implemented.');
   }
@@ -137,7 +137,7 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   /// should implement an on-disk cache.
   Future<void> clearTileCache(
     TileOverlayId tileOverlayId, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('clearTileCache() has not been implemented.');
   }
@@ -148,7 +148,7 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   /// platform side.
   Future<void> animateCamera(
     CameraUpdate cameraUpdate, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('animateCamera() has not been implemented.');
   }
@@ -159,7 +159,7 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   /// platform side.
   Future<void> moveCamera(
     CameraUpdate cameraUpdate, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('moveCamera() has not been implemented.');
   }
@@ -176,14 +176,14 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   /// The style string can be generated using [map style tool](https://mapstyle.withgoogle.com/).
   Future<void> setMapStyle(
     String mapStyle, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('setMapStyle() has not been implemented.');
   }
 
   /// Return the region that is visible in a map.
   Future<LatLngBounds> getVisibleRegion({
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('getVisibleRegion() has not been implemented.');
   }
@@ -195,7 +195,7 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   /// of the map, not necessarily of the whole screen.
   Future<ScreenCoordinate> getScreenCoordinate(
     LatLng latLng, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('getScreenCoordinate() has not been implemented.');
   }
@@ -207,7 +207,7 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   /// of the map, not necessarily of the whole screen.
   Future<LatLng> getLatLng(
     ScreenCoordinate screenCoordinate, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('getLatLng() has not been implemented.');
   }
@@ -222,7 +222,7 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   ///   * [isMarkerInfoWindowShown] to check if the Info Window is showing.
   Future<void> showMarkerInfoWindow(
     MarkerId markerId, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError(
         'showMarkerInfoWindow() has not been implemented.');
@@ -238,7 +238,7 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   ///   * [isMarkerInfoWindowShown] to check if the Info Window is showing.
   Future<void> hideMarkerInfoWindow(
     MarkerId markerId, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError(
         'hideMarkerInfoWindow() has not been implemented.');
@@ -252,23 +252,23 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   /// * See also:
   ///   * [showMarkerInfoWindow] to show the Info Window.
   ///   * [hideMarkerInfoWindow] to hide the Info Window.
-  Future<bool> isMarkerInfoWindowShown(
+  Future<bool?> isMarkerInfoWindowShown(
     MarkerId markerId, {
-    @required int mapId,
+    required int mapId,
   }) {
     throw UnimplementedError('updateMapOptions() has not been implemented.');
   }
 
   /// Returns the current zoom level of the map
-  Future<double> getZoomLevel({
-    @required int mapId,
+  Future<double?> getZoomLevel({
+    required int mapId,
   }) {
     throw UnimplementedError('getZoomLevel() has not been implemented.');
   }
 
   /// Returns the image bytes of the map
-  Future<Uint8List> takeSnapshot({
-    @required int mapId,
+  Future<Uint8List?> takeSnapshot({
+    required int mapId,
   }) {
     throw UnimplementedError('takeSnapshot() has not been implemented.');
   }
@@ -277,62 +277,62 @@ abstract class GoogleMapsFlutterPlatform extends PlatformInterface {
   // into the plugin
 
   /// The Camera started moving.
-  Stream<CameraMoveStartedEvent> onCameraMoveStarted({@required int mapId}) {
+  Stream<CameraMoveStartedEvent> onCameraMoveStarted({required int mapId}) {
     throw UnimplementedError('onCameraMoveStarted() has not been implemented.');
   }
 
   /// The Camera finished moving to a new [CameraPosition].
-  Stream<CameraMoveEvent> onCameraMove({@required int mapId}) {
+  Stream<CameraMoveEvent> onCameraMove({required int mapId}) {
     throw UnimplementedError('onCameraMove() has not been implemented.');
   }
 
   /// The Camera is now idle.
-  Stream<CameraIdleEvent> onCameraIdle({@required int mapId}) {
+  Stream<CameraIdleEvent> onCameraIdle({required int mapId}) {
     throw UnimplementedError('onCameraMove() has not been implemented.');
   }
 
   /// A [Marker] has been tapped.
-  Stream<MarkerTapEvent> onMarkerTap({@required int mapId}) {
+  Stream<MarkerTapEvent> onMarkerTap({required int mapId}) {
     throw UnimplementedError('onMarkerTap() has not been implemented.');
   }
 
   /// An [InfoWindow] has been tapped.
-  Stream<InfoWindowTapEvent> onInfoWindowTap({@required int mapId}) {
+  Stream<InfoWindowTapEvent> onInfoWindowTap({required int mapId}) {
     throw UnimplementedError('onInfoWindowTap() has not been implemented.');
   }
 
   /// A [Marker] has been dragged to a different [LatLng] position.
-  Stream<MarkerDragEndEvent> onMarkerDragEnd({@required int mapId}) {
+  Stream<MarkerDragEndEvent> onMarkerDragEnd({required int mapId}) {
     throw UnimplementedError('onMarkerDragEnd() has not been implemented.');
   }
 
   /// A [Polyline] has been tapped.
-  Stream<PolylineTapEvent> onPolylineTap({@required int mapId}) {
+  Stream<PolylineTapEvent> onPolylineTap({required int mapId}) {
     throw UnimplementedError('onPolylineTap() has not been implemented.');
   }
 
   /// A [Polygon] has been tapped.
-  Stream<PolygonTapEvent> onPolygonTap({@required int mapId}) {
+  Stream<PolygonTapEvent> onPolygonTap({required int mapId}) {
     throw UnimplementedError('onPolygonTap() has not been implemented.');
   }
 
   /// A [Circle] has been tapped.
-  Stream<CircleTapEvent> onCircleTap({@required int mapId}) {
+  Stream<CircleTapEvent> onCircleTap({required int mapId}) {
     throw UnimplementedError('onCircleTap() has not been implemented.');
   }
 
   /// A Map has been tapped at a certain [LatLng].
-  Stream<MapTapEvent> onTap({@required int mapId}) {
+  Stream<MapTapEvent> onTap({required int mapId}) {
     throw UnimplementedError('onTap() has not been implemented.');
   }
 
   /// A Map has been long-pressed at a certain [LatLng].
-  Stream<MapLongPressEvent> onLongPress({@required int mapId}) {
+  Stream<MapLongPressEvent> onLongPress({required int mapId}) {
     throw UnimplementedError('onLongPress() has not been implemented.');
   }
 
   /// Dispose of whatever resources the `mapId` is holding on to.
-  void dispose({@required int mapId}) {
+  void dispose({required int mapId}) {
     throw UnimplementedError('dispose() has not been implemented.');
   }
 
