@@ -12,16 +12,16 @@ class Tile {
   const Tile(this.width, this.height, this.data);
 
   /// The width of the image encoded by data in logical pixels.
-  final int width;
+  final int? width;
 
   /// The height of the image encoded by data in logical pixels.
-  final int height;
+  final int? height;
 
   /// A byte array containing the image data.
   ///
   /// The image data format must be natively supported for decoding by the platform.
   /// e.g on Android it can only be one of the [supported image formats for decoding](https://developer.android.com/guide/topics/media/media-formats#image-formats).
-  final Uint8List data;
+  final Uint8List? data;
 
   /// Converts this object to JSON.
   Map<String, dynamic> toJson() {
