@@ -77,7 +77,7 @@ class BitmapDescriptor {
   /// Use [fromAssetImage]. This method does not respect the screen dpi when
   /// picking an asset image.
   @Deprecated("Use fromAssetImage instead")
-  static BitmapDescriptor fromAsset(String assetName, {String package}) {
+  static BitmapDescriptor fromAsset(String assetName, {String? package}) {
     if (package == null) {
       return BitmapDescriptor._(<dynamic>[_fromAsset, assetName]);
     } else {
@@ -95,8 +95,8 @@ class BitmapDescriptor {
   static Future<BitmapDescriptor> fromAssetImage(
     ImageConfiguration configuration,
     String assetName, {
-    AssetBundle bundle,
-    String package,
+    AssetBundle? bundle,
+    String? package,
     bool mipmaps = true,
   }) async {
     if (!mipmaps && configuration.devicePixelRatio != null) {
@@ -114,10 +114,10 @@ class BitmapDescriptor {
       _fromAssetImage,
       assetBundleImageKey.name,
       assetBundleImageKey.scale,
-      if (kIsWeb && configuration?.size != null)
+      if (kIsWeb && configuration.size != null)
         [
-          configuration.size.width,
-          configuration.size.height,
+          configuration.size!.width,
+          configuration.size!.height,
         ],
     ]);
   }
