@@ -31,7 +31,7 @@ class MarkerUpdates {
     final Set<MarkerId> currentMarkerIds = currentMarkers.keys.toSet();
 
     Marker idToCurrentMarker(MarkerId id) {
-      return currentMarkers[id];
+      return currentMarkers[id]!;
     }
 
     final Set<MarkerId> _markerIdsToRemove =
@@ -45,7 +45,7 @@ class MarkerUpdates {
     /// Returns `true` if [current] is not equals to previous one with the
     /// same id.
     bool hasChanged(Marker current) {
-      final Marker previous = previousMarkers[current.markerId];
+      final Marker previous = previousMarkers[current.markerId]!;
       return current != previous;
     }
 
@@ -61,13 +61,13 @@ class MarkerUpdates {
   }
 
   /// Set of Markers to be added in this update.
-  Set<Marker> markersToAdd;
+  Set<Marker>? markersToAdd;
 
   /// Set of MarkerIds to be removed in this update.
-  Set<MarkerId> markerIdsToRemove;
+  Set<MarkerId>? markerIdsToRemove;
 
   /// Set of Markers to be changed in this update.
-  Set<Marker> markersToChange;
+  Set<Marker>? markersToChange;
 
   /// Converts this object to something serializable in JSON.
   Map<String, dynamic> toJson() {
@@ -79,10 +79,10 @@ class MarkerUpdates {
       }
     }
 
-    addIfNonNull('markersToAdd', serializeMarkerSet(markersToAdd));
-    addIfNonNull('markersToChange', serializeMarkerSet(markersToChange));
+    addIfNonNull('markersToAdd', serializeMarkerSet(markersToAdd!));
+    addIfNonNull('markersToChange', serializeMarkerSet(markersToChange!));
     addIfNonNull('markerIdsToRemove',
-        markerIdsToRemove.map<dynamic>((MarkerId m) => m.value).toList());
+        markerIdsToRemove!.map<dynamic>((MarkerId m) => m.value).toList());
 
     return updateMap;
   }
@@ -91,7 +91,7 @@ class MarkerUpdates {
   bool operator ==(Object other) {
     if (identical(this, other)) return true;
     if (other.runtimeType != runtimeType) return false;
-    final MarkerUpdates typedOther = other;
+    final MarkerUpdates typedOther = other as MarkerUpdates;
     return setEquals(markersToAdd, typedOther.markersToAdd) &&
         setEquals(markerIdsToRemove, typedOther.markerIdsToRemove) &&
         setEquals(markersToChange, typedOther.markersToChange);
