@@ -14,7 +14,7 @@ Map<CircleId, Circle> keyByCircleId(Iterable<Circle> circles) {
 }
 
 /// Converts a Set of Circles into something serializable in JSON.
-List<Map<String, dynamic>> serializeCircleSet(Set<Circle> circles) {
+List<Map<String, dynamic>>? serializeCircleSet(Set<Circle> circles) {
   if (circles == null) {
     return null;
   }
