@@ -14,7 +14,7 @@ Map<PolygonId, Polygon> keyByPolygonId(Iterable<Polygon> polygons) {
 }
 
 /// Converts a Set of Polygons into something serializable in JSON.
-List<Map<String, dynamic>> serializePolygonSet(Set<Polygon> polygons) {
+List<Map<String, dynamic>>? serializePolygonSet(Set<Polygon> polygons) {
   if (polygons == null) {
     return null;
   }
