@@ -12,7 +12,7 @@ Map<TileOverlayId, TileOverlay> keyTileOverlayId(
 }
 
 /// Converts a Set of TileOverlays into something serializable in JSON.
-List<Map<String, dynamic>> serializeTileOverlaySet(
+List<Map<String, dynamic>>? serializeTileOverlaySet(
     Set<TileOverlay> tileOverlays) {
   if (tileOverlays == null) {
     return null;
